src/server/requirements.txt
pyserial
mariadb
